Add files via upload
This program details steps on how to use a walmart receipt and extract the text data. Obviously i have not given my receipt for security reasons. However the steps are the same if you need to use this program as an example for your own receipt. Please note that the result of this script also depends on the image quality of the receipt. I will try to use another receipt that is directly downloaded from the walmart link that is downloaded when we choose the option of sending the receipt to our phone no. when we pay at the kiosk or the Point of Sale (Pos)