Add files via upload
This is to process receipts that are downloaded from walmart website or your walmart account. For every purchase, the scanned copy of the bill is uploaded. If you think you do not want to go through the process of manually scanning the receipt and having to deal with the preprocessing of image, this would be ideal. Atleast the format and the quality of the print/receipt is going to be uniform.